Changed DragBox stop events (non-first) to click
DragBox used to handle mouseup events exclusively. This is not very pretty: after the user lets go of the mouse button the first time, it should handle clicks instead.

This change fixes the bug in which nodes created using two or more clicks are not selected and editable immediately after creation. Unfortunately, the fix is "clever": it works because the click handler on the treeDiv, which deselects the selected element (if any), fires earlier than the click handler on the document, which now finishes the dragging action and creates the element. If you find a cleaner solution (one that also does not involve additional state), feel free to implement it.