Switched to direct CSS position retrieval in Save
Save used to use .position() to retrieve nodes' positions, which has a serious disadvantage: its results are not reliable when the element is hidden, which is always the case for all non-zoomed trees (even with the show/hide trick). Retrieving the left and top CSS values using .css() *does* work for hidden elements. The downside is that it will fail when different positioning methods, or even different CSS units, are applied to nodes. For now we know the jsPlumb (specifically Katavorio) dragging implementation will always use pixels, which suffices.