Turned off minimap by default
The minimap is an expensive feature: it leads to a noticeable performance degradation when editing large scenarios. Dragging performance, in particular, is "passable" in Chrome, "barely acceptable" in Firefox, and "awful" in Internet Explorer.

When users, especially new users, experience performance problems, they're unlikely to correctly recognise that the minimap is the culprit and that it can be disabled manually. Because first impressions are important, we've decided to turn off the minimap by default. This should be re-evaluated once overall performance improvements have been made and the minimap UI has been improved.