Simplified grid initialisation
The purpose of initialiseGrid was apparently twofold:

1. setting gridX and gridY based on the actual (pixel-based) width and height of treeContainer of the first tree that is created
2. adjusting gridX and gridY to avoid rounding problems

The latter of these was only necessary when the width of treeContainers was specified as a percentage, which is no longer the case. The former is smart in that it uses the actual size, but problematic because it is unreliable when the treeContainer is not being displayed.

The new method looks up the size in the style sheet, which can be done at any time and can be adjusted to use percentages too.