Added JSHint configuration
The proper way to use globals is using the inline "globals" directive. The tools (editors and IDEs) we use do not have support for making the management and maintenance of those directives feasible, so for now, we'll use a global "predef" directive instead. The globals are listed in "import" (script tag) order.